Switched Build To Browserify
Switched build process to Browserify in order to allow packaging multiple js files together into a single package.  The Tinyify plugin is used for the packing process.  JQuery is now packed into a single distribution with the main zchat.js code as dist/zchat.min.js.  This will be tracked for now and tagged for appropriate versions in the future.  Added the ability to click on images and view them full-screen in a lightbox.  The lightbox also includes a caption of the image which can be clicked to copy the link explicitly.  Modified css to include this modal component and sub-components as well as their animations.  Currently, the zoom animation is not working as planned, but a future update should fix this.